src/lambda/po-workflow/requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
boto3==1.38.17
botocore==1.38.17
brotli==1.1.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
comm==0.2.2
cssselect2==0.8.0
debugpy==1.8.14
decorator==5.2.1
events==0.5
executing==2.2.0
fonttools==4.58.5
ghostscript==0.8.1
gitdb==4.0.12
gitpython==3.1.44
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==9.2.0
ipython-pygments-lexers==1.1.1
jedi==0.19.2
jinja2==3.1.6
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter-client==8.6.3
jupyter-core==5.7.2
langchain==0.3.25
langchain-core==0.3.60
langchain-text-splitters==0.3.8
langgraph==0.4.10
langgraph-checkpoint==2.1.0
langgraph-prebuilt==0.2.2
langgraph-sdk==0.1.70
langsmith==0.3.42
markdown2==2.5.3
markupsafe==3.0.2
matplotlib-inline==0.1.7
narwhals==1.39.1
nest-asyncio==1.6.0
numpy==2.2.5
opensearch-py==3.0.0
orjson==3.10.18
ormsgpack==1.10.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
prompt-toolkit==3.0.51
protobuf==6.31.0
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
pyarrow==20.0.0
pycparser==2.22
pydantic==2.11.4
pydantic-core==2.33.2
pydeck==0.9.1
pydyf==0.11.0
pygments==2.19.1
pypdf2==3.0.1
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pyyaml==6.0.2
pyzmq==26.4.0
referencing==0.36.2
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.25.0
s3transfer==0.12.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sqlalchemy==2.0.41
stack-data==0.6.3
streamlit==1.45.1
tenacity==9.1.2
tinycss2==1.4.0
tinyhtml5==2.0.0
toml==0.10.2
tornado==6.4.2
traitlets==5.14.3
typing-extensions==4.13.2
typing-inspection==0.4.0
tzdata==2025.2
urllib3==1.26.20
watchdog==6.0.0
wcwidth==0.2.13
weasyprint<53
webencodings==0.5.1
xxhash==3.5.0
zopfli==0.2.3.post1
zstandard==0.23.0

